backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
gunicorn>=21.2.0
pydantic>=2.5.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.1
redis>=5.0.1
celery>=5.3.6
httpx>=0.25.2
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
numpy>=1.26.2
scipy>=1.11.4
scikit-learn>=1.3.2
spaCy>=3.7.2
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic>=0.4.27
minio>=7.2.0
sentence-transformers>=2.2.2
openai>=1.12.0
python-dotenv>=1.0.0
hypothesis>=6.88.0
kafka-python>=2.0.2
pymupdf>=1.23.27
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.26.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
feedparser>=6.0.10
prometheus-client>=0.19.0
opentelemetry-distro>=0.43b0
opentelemetry-instrumentation>=0.43b0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-exporter-jaeger>=1.20.0
cryptography>=41.0.0
hvac>=2.1.0
authlib>=1.3.0
httpx>=0.25.2
